fix tolerance datatype
Former-commit-id: e79baaeb7a8f042f9e75e777f73740f37ef8f4e2

+++ test/common/TestRunner.cc
@@ -33,7 +33,7 @@ int run_sequence(const std::string& model,
                  const std::vector<std::string>& tests,
                  const std::vector<std::string>& sims,
                  const std::string& data_dir, const std::string& out_dir,
-                 const int tolerance) {
+                 const float tolerance) {
   // Set data parameters
   bool use_data_file = true;
   std::map<std::string, MemoryMap> mem_map;
